+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Core library changes
 
+* `dart:io`
+  * Added `Platform.localeName`.
+
 ### Dart VM
 
 ### Strong Mode